ai-service/requirements.txt
flask
Flask-Cors
pymongo
scikit-learn
numpy
python-dotenv
pandas
